+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - move some models into common namespace
 - support for all `release` events
 - support for all `issue_comment` events
+- support for all `repository` events
 
 ## 1.1.2 - 2024-07-26